requirements.txt
groq
python-dotenv
opencv-python
pillow
google-generativeai
pyperclip
elevenlabs
duckduckgo-search  
selenium
webdriver-manager